Initial commit
This repo is for the first deliverable task of the second block of the
'HPC Tools' subject. In this repo, I'll implement a dense matrix
multiplication, and compare it's time performance agains openBLAS'
cblas_dgemv function. Also, I'll convert the dense matrix to a sparse
matrix in CSR format, and again compare it's time performance against
the CSPARSE library.